develop
카드 수 & 패스 수 입력 받는 화면과 카드 뽑는 화면 분리
