+++ src/components/css/card.css
@@ -29,4 +29,4 @@
   -webkit-transform: translateY(-20px);
           transform: translateY(-20px);
 }
-/*# sourceMappingURL=Card.css.map */
+/*# sourceMappingURL=RoletteCard.css.map */
