+++ src/views/RouletteGameView.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import MyInput from '../components/MyInput';
+import RouletteInput from '../components/RouletteInput';
 
 export default function RouletteGameView() {
   return (
     <>
       <div>
-        <MyInput/>
+        <RouletteInput/>
       </div>
     </>
   );
